增加查询交易详情接口

+++ client/commands/get_tx.go
@@ -33,7 +33,7 @@ import (
 	"time"
 )
 
-func GetTransactionJson(client *jsonrpc.BasicClient, chainId int, txHash *txprotocal.NulsHash) (string, error) {
+func GetTransactionJson(client *jsonrpc.BasicClient, chainId uint16, txHash *txprotocal.NulsHash) (string, error) {
 	if client == nil || txHash == nil {
 		return "", errors.New("parameter wrong.")
 	}
